+++ iac/pulumi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cf-eventhub/pulumi
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [c3867ae]
+- Updated dependencies [54f2bed]
+  - cf-eventhub@0.2.0
+
 ## 0.0.19
 
 ### Patch Changes